refactor: restructure components folder
