+++ src/components/schedule/class/ClassInfoEntry.tsx
