+++ src/components/user/settings/CalendarFeed.tsx
@@ -8,7 +8,7 @@ import TextField from "@mui/material/TextField";
 import Tooltip from "@mui/material/Tooltip";
 import { useState } from "react";
 
-import SubHeader from "@/components/modals/SubHeader";
+import SubHeader from "@/components/utils/SubHeader";
 import { useCopyToClipboard } from "@/lib/hooks/useCopyToClipboard";
 import { useUserCalendarFeedUrl } from "@/lib/hooks/useUserCalendarFeedUrl";
