+++ src/routes/login.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 
-import AutoLogin from "@/components/AutoLogin";
+import AutoLogin from "@/components/user/AutoLogin";
 
 export const Route = createFileRoute("/login")({
     component: AutoLogin,
